reorganizes files & packages
