+++ app/src/main/java/angelosz/catbattlegame/data/repository/OfflinePlayerAccountRepository.kt
@@ -6,7 +6,7 @@ import angelosz.catbattlegame.data.entities.PlayerAccount
 import angelosz.catbattlegame.data.entities.PlayerTeam
 import angelosz.catbattlegame.data.entities.PlayerTeamOwnedCat
 import angelosz.catbattlegame.domain.models.BasicCatData
-import angelosz.catbattlegame.ui.armory.cats_view.SimpleArmoryCatData
+import angelosz.catbattlegame.ui.armory.data.SimpleArmoryCatData
 import kotlinx.coroutines.flow.Flow
 
 class OfflinePlayerAccountRepository(val dao: PlayerDao): PlayerAccountRepository {
