+++ app/src/main/java/angelosz/catbattlegame/ui/armory/cats_view/ArmoryCatsUiState.kt
@@ -1,6 +1,8 @@
 package angelosz.catbattlegame.ui.armory.cats_view
 
 import angelosz.catbattlegame.domain.enums.ScreenState
+import angelosz.catbattlegame.ui.armory.data.DetailedArmoryCatData
+import angelosz.catbattlegame.ui.armory.data.SimpleArmoryCatData
 
 data class ArmoryCatsUiState (
     val state: ScreenState = ScreenState.INITIALIZING,
